+++ tools/xdhq/examples/JRE/Hello/Hello.java
@@ -42,7 +42,7 @@ class Hello extends Atlas {
 	{
 		switch( action) {
 		case "":
-			dom.setLayout("", html);
+			dom.setLayout("", html );
 			dom.setContent( "input", "World");
 			break;
 		case "Submit":
@@ -60,7 +60,9 @@ class Hello extends Atlas {
 	}
 
 	public static void main(String[] args) throws Exception {
-		String head = "<title>\"Hello, World !\" example</title>";
+		String head =
+			"<title>\"Hello, World !\" example</title>" +
+			"<link rel=\"icon\" type=\"image/png\" href=\"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAgBAMAAACBVGfHAAAAMFBMVEUEAvyEhsxERuS8urQsKuycnsRkYtzc2qwUFvRUVtysrrx0ctTs6qTMyrSUksQ0NuyciPBdAAABHklEQVR42mNgwAa8zlxjDd2A4POfOXPmzZkFCAH2M8fNzyALzDlzg2ENssCbMwkMOsgCa858YOjBKxBzRoHhD7LAHiBH5swCT9HQ6A9ggZ4zp7YCrV0DdM6pBpAAG5Blc2aBDZA68wCsZPuZU0BDH07xvHOmAGKKvgMP2NA/Zw7ADIYJXGDgLQeBBSCBFu0aoAPYQUadMQAJAE29zwAVWMCWpgB08ZnDQGsbGhpsgCqBQHNfzRkDEIPlzFmo0T5nzoMovjPHoAK8Zw5BnA5yDosDSAVYQOYMKIDZzkoDzagAsjhqzjRAfXTmzAQgi/vMQZA6pjtAvhEk0E+ATWRRm6YBZuScCUCNN5szH1D4TGdOoSrggtiNAH3vBBjwAQCglIrSZkf1MQAAAABJRU5ErkJggg==\" />";
 
 		launch(() -> new Hello(), head, "Hello", GUI.DEFAULT, args );
 	}