+++ tools/xdhq/examples/RBY/Hello/main.rb
@@ -40,7 +40,6 @@ def acClear(userObject, dom, id)
 	dom.focus("input")
 end
 
-
 callbacks = {
 	"" => method(:acConnect),
 	"Submit" => method(:acSubmit),