+++ tools/xdhq/examples/ZND/Hello/Hello.php
@@ -50,20 +50,14 @@ getAtlas();
 class Hello extends Threaded {
  // The content of the 'Main.html' file.
  static $body = <<<EOT
-<div class="vcenter-out">
- <div class="vcenter-in">
-  <fieldset>
-   <label>Name:</label>
-   <input id="input" maxlength="20" placeholder="Enter a name here" type="text" data-xdh-onevent="input|Typing" />
+<div style="display: table; margin: 50px auto auto auto;">
+ <fieldset>
+  <input id="input" maxlength="20" placeholder="Enter a name here" type="text" data-xdh-onevent="Submit" value="World"/>
+  <div style="display: flex; justify-content: space-around; margin: 5px auto auto auto;">
+   <button data-xdh-onevent="Submit">Submit</button>
    <button data-xdh-onevent="Clear">Clear</button>
-   <hr />
-   <h1>
-    <span>Hello </span>
-    <span style="font-style: italic;" id="name"></span>
-    <span>!</span>
-   </h1>
-  </fieldset>
- </div>
+  </div>
+ </fieldset>
 </div>
 EOT;
 
@@ -73,19 +67,22 @@ EOT;
    $dom->setLayout("", self::$body);
    $dom->focus("input");
    break;
-  case "Typing":
-   $dom->setContent("name", $dom->getContent($id));
+  case "Submit":
+   $dom->alert("Hello, " . $dom->getContent("input") . "!" );
+   $dom->focus("input");
    break;
   case "Clear":
    if ($dom->confirm("Are you sure?")) {
-    $dom->setContents(["input" => "", "name" => ""]);
+    $dom->setContent("input", "");
    }
+   $dom->focus("input");
    break;
   default:
    throw new Exception("Unknown action '" . $action . "' !!!");
   }
  }
 }
+
 
 function hello() {
  return new Hello();
@@ -95,26 +92,6 @@ function hello() {
 $head = <<<EOT
 <title>"Hello, World !" example</title>
 <link rel="icon" type="image/png" href="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAgBAMAAACBVGfHAAAAMFBMVEUEAvyEhsxERuS8urQsKuycnsRkYtzc2qwUFvRUVtysrrx0ctTs6qTMyrSUksQ0NuyciPBdAAABHklEQVR42mNgwAa8zlxjDd2A4POfOXPmzZkFCAH2M8fNzyALzDlzg2ENssCbMwkMOsgCa858YOjBKxBzRoHhD7LAHiBH5swCT9HQ6A9ggZ4zp7YCrV0DdM6pBpAAG5Blc2aBDZA68wCsZPuZU0BDH07xvHOmAGKKvgMP2NA/Zw7ADIYJXGDgLQeBBSCBFu0aoAPYQUadMQAJAE29zwAVWMCWpgB08ZnDQGsbGhpsgCqBQHNfzRkDEIPlzFmo0T5nzoMovjPHoAK8Zw5BnA5yDosDSAVYQOYMKIDZzkoDzagAsjhqzjRAfXTmzAQgi/vMQZA6pjtAvhEk0E+ATWRRm6YBZuScCUCNN5szH1D4TGdOoSrggtiNAH3vBBjwAQCglIrSZkf1MQAAAABJRU5ErkJggg==" />
-<style type="text/css">
- html,
- body {
-  height: 100%;
-  padding: 0;
-  margin: 0;
- }
-
- .vcenter-out,
- .hcenter {
-  display: table;
-  height: 100%;
-  margin: auto;
- }
-
- .vcenter-in {
-  display: table-cell;
-  vertical-align: middle;
- }
-</style>
 EOT;
 
 Atlas::launch('hello', $head, null, "Hello");