+++ tools/xdhq/wrappers/JRE/DOM.java
@@ -64,7 +64,9 @@ public class DOM {
 	}
 
 	public void alert(String message) {
-		DOM.call("Alert_1", Type.VOID, new String[] { message }, emptys);
+		DOM.call("Alert_1", Type.STRING, new String[] { message }, emptys);
+		// For the return value being 'STRING' instead of 'VOID',
+		// see the 'alert' primitive in 'XDHqXDH'.
 	}
 
 	public boolean confirm(String message) {